added toggleNotification to repository

+++ backend/superprice/src/main/java/com/example/superprice/model/Product.java
@@ -1,6 +1,6 @@
 package com.example.superprice.model;
 
-public record Product(Long id, String name, String description, String store,
+public record Product(int id, String name, String description, String store,
                       String imageURL, double price, int quantity, double discount, boolean notification,
                       String currDate, String promoStartDate, String promoEndDate) {
 }
